Partially refactored acceptance tests
* Expression
* Functions
